+++ policyengine_core/enums/enum.py
@@ -53,10 +53,11 @@ class Enum(enum.Enum):
         # Confusingly, Numpy uses "S" to refer to byte-string arrays
         # and "U" to refer to Unicode-string arrays, which are also
         # referred to as the "str" type
-        if array.dtype.kind == "S":
+        if isinstance(array[0], Enum):
+            array = np.array([item.name for item in array])
+        if array.dtype.kind == "S" or array.dtype == object:
             # Convert boolean array to string array
             array = array.astype(str)
-
         if isinstance(array, np.ndarray) and array.dtype.kind in {"U", "S"}:
             # String array
             indices = np.select(