+++ policyengine_core/tracers/__init__.py
@@ -2,6 +2,7 @@ from .computation_log import ComputationLog
 from .flat_trace import FlatTrace
 from .full_tracer import FullTracer
 from .performance_log import PerformanceLog
+from .variable_graph import VariableGraph
 from .simple_tracer import SimpleTracer
 from .trace_node import TraceNode
 from .tracing_parameter_node_at_instant import TracingParameterNodeAtInstant